+++ js/main.js
@@ -2,9 +2,5 @@
  * Created by Denis on 27.03.2018.
  */
 
-import {Application} from './application';
-import {saveResults, getStats} from './modules/game';
-// saveResults({points:1});
-
-// Application.showStats();
+import {Application} from './common/application';
 Application.start();
